Moved deprecated code out of BrowserMobProxyServer into BrowserMobProxyServerLegacyAdapter. BrowserMobProxyServer no longer implements LegacyProxyServer.
