+++ browsermob-rest/src/main/java/net/lightbody/bmp/proxy/guice/LegacyProxyServerProvider.java
@@ -2,6 +2,7 @@ package net.lightbody.bmp.proxy.guice;
 
 import com.google.inject.Provider;
 import net.lightbody.bmp.BrowserMobProxyServer;
+import net.lightbody.bmp.BrowserMobProxyServerLegacyAdapter;
 import net.lightbody.bmp.proxy.LegacyProxyServer;
 import net.lightbody.bmp.proxy.ProxyServer;
 
@@ -12,7 +13,7 @@ public class LegacyProxyServerProvider implements Provider<LegacyProxyServer> {
     @Override
     public LegacyProxyServer get() {
         if (useLittleProxy || Boolean.getBoolean("bmp.use.littleproxy")) {
-            return new BrowserMobProxyServer();
+            return new BrowserMobProxyServerLegacyAdapter();
         } else {
             return new ProxyServer();
         }
